+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-17T12:17:48.632Z",
+    "2026-07-20T03:50:54.682Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {