+++ packages/orq-rc/src/models/operations/updatetool.ts
@@ -1472,7 +1472,7 @@ export const RequestBodyTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RequestBodyTools
 > = z.object({
-  id: z.string().default("01KXR06R75M3F8X40MTNMRQ2SC"),
+  id: z.string().default("01KXYTCVJDVB4XN3HV8CTKDY8Q"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => UpdateToolRequestBodyToolsSchema$outboundSchema),
@@ -2193,7 +2193,7 @@ export const UpdateToolResponseBodyCodeExecutionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KXR06R6VZWSZJFGV42H5Q1Q2"),
+  _id: z.string().default("tool_01KXYTCVJ56WFP00ECS7JMQMX0"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2296,7 +2296,7 @@ export const UpdateToolResponseBodyTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KXR06R6TGGPN8KC7SPWH1X1K"),
+  id: z.string().default("01KXYTCVJ4MY407D6475PWRSQ8"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => UpdateToolResponseBodyToolsSchema$inboundSchema),
@@ -2354,7 +2354,7 @@ export const UpdateToolResponseBodyMCPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KXR06R6SS6YWJ2MGBYBSEH3P"),
+  _id: z.string().default("tool_01KXYTCVJ2WDFTWWVQNJQHQXEG"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2543,7 +2543,7 @@ export const UpdateToolResponseBodyHTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KXR06R6P1ZDPX1X90Y6CYK7B"),
+  _id: z.string().default("tool_01KXYTCVHZY5PAVG7F9RAZH9BE"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2638,7 +2638,7 @@ export const UpdateToolResponseBodyJSONSchemaTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KXR06R6MBKGBPPX5YRC5AFA2"),
+  _id: z.string().default("tool_01KXYTCVHXAGF7T1C3DK6W2DR6"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2741,7 +2741,7 @@ export const UpdateToolResponseBodyFunctionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KXR06R6J8KPVMYT84J24VSGR"),
+  _id: z.string().default("tool_01KXYTCVHV183HNTMQMG8J7609"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),