+++ packages/orq-rc/RELEASES.md
@@ -10919,3 +10919,13 @@ Based on:
 - [typescript v4.12.0-rc.38] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.38 - packages/orq-rc
+
+## 2026-07-20 03:53:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.2 (2.918.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.39] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.39 - packages/orq-rc